libexec/cli/build.exec: Added perms corrections to docker OLD_EXTRACTION case
This is a simple fix for the OLD_EXTRACTION scenario for docker images.

For builds from multi-layer docker images, early layers can place
read-only directories that later layers try to write to.  Unless the
build is run as root, these permissions are respected, and the build
fails.  This commit changes the functionality to make all read-only
directories temporarily writeable for untarring into.